# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Layer": {
        "enum": [
          "territories",
          "fire_events",
          "observations",
          "lightning",
          "risk",
          "parcels"
        ],
        "title": "Layer",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Fire, spraying conditions and field anomalies per field and section.",
    "title": "Field Watch API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tiles/{layer}/{z}/{x}/{y}.pbf": {
      "get": {
        "operationId": "tile_tiles__layer___z___x___y__pbf_get",
        "parameters": [
          {
            "in": "path",
            "name": "layer",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Layer"
            }
          },
          {
            "in": "path",
            "name": "z",
            "required": true,
            "schema": {
              "maximum": 22,
              "minimum": 0,
              "title": "Z",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "x",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "X",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "y",
            "required": true,
            "schema": {
              "minimum": 0,
              "title": "Y",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tile",
        "tags": [
          "tiles"
        ]
      }
    }
  }
}